+++ assets/js/disable-blog-customizer.js
@@ -1,0 +1,13 @@
+(function() {
+	'use strict';
+
+	wp.customize.bind( 'ready', function() {
+
+		// Replace the default text in the homepage settings with the new version.
+		if ( typeof dwpbCustomizer.homepageSettingsText !== 'undefined' ) {
+			wp.customize.section("static_front_page").container.find(".customize-section-description")[0].innerText = dwpbCustomizer.homepageSettingsText;
+		} // end if
+
+	} );
+
+})();