+++ disable-blog.php
@@ -15,7 +15,7 @@
  * Plugin Name: Disable Blog
  * Plugin URI:  https://wordpress.org/plugins/disable-blog/
  * Description: Go blog-less with WordPress. This plugin disables all blog-related functionality (by hiding, removing, and redirecting).
- * Version:     0.5.0
+ * Version:     0.5.1
  * Author:      Joshua Nelson
  * Author URI:  http://joshuadnelson.com
  * License:     GPL-2.0+
@@ -60,9 +60,14 @@ function deactivate_disable_blog() {
 register_activation_hook( __FILE__, 'activate_disable_blog' );
 register_deactivation_hook( __FILE__, 'deactivate_disable_blog' );
 
+// Constants.
+define( 'DWPB_DIR', dirname( __FILE__ ) );
+define( 'DWPB_URL', plugins_url( '/', __FILE__ ) );
+define( 'DWPB_PLUGIN_NAME', 'disable-blog' );
+define( 'DWPB_VERSION', '0.5.1' );
+
 /**
- * The core plugin class that is used to define internationalization,
- * admin-specific hooks, and public-facing site hooks.
+ * The core plugin class that is used to define everything.
  */
 require plugin_dir_path( __FILE__ ) . 'includes/class-disable-blog.php';
 
@@ -76,7 +81,7 @@ require plugin_dir_path( __FILE__ ) . 'includes/class-disable-blog.php';
  * @since 0.4.0
  */
 function run_disable_blog() {
-	$plugin = new Disable_Blog();
+	$plugin = new Disable_Blog( DWPB_PLUGIN_NAME, DWPB_VERSION );
 	$plugin->run();
 }
 add_action( 'plugins_loaded', 'run_disable_blog', 10, 0 );