+++ includes/class-disable-blog-functions.php
@@ -11,10 +11,7 @@
 /**
  * Main class for functions used by other classes.
  *
- * @since      0.5.0
- * @package    Disable_Blog
- * @subpackage Disable_Blog_Functions
- * @author     Joshua Nelson <josh@joshuadnelson.com>
+ * @since 0.5.0
  */
 class Disable_Blog_Functions {
 
@@ -245,6 +242,27 @@ class Disable_Blog_Functions {
 		 * @return bool
 		 */
 		return (bool) apply_filters( 'dwpb_disable_blog', false );
+
+	}
+
+	/**
+	 * Toggle the disable feed via this filter.
+	 *
+	 * @since 0.5.1
+	 * @param object $post            Global post object.
+	 * @param bool   $is_comment_feed True if the feed is a comment feed.
+	 */
+	public function disable_feeds( $post, $is_comment_feed = false ) {
+
+		/**
+		 * Toggle the disable feed via this filter.
+		 *
+		 * @since 0.4.0
+		 * @param bool $bool True to cancel the feed, assuming it's a post feed.
+		 * @param object $post Global post object.
+		 * @param bool $is_comment_feed True if the feed is a comment feed.
+		 */
+		return (bool) apply_filters( 'dwpb_disable_feed', true, $post, $is_comment_feed );
 
 	}