+++ includes/class-disable-blog-loader.php
@@ -4,9 +4,9 @@
  *
  * @link       https://github.com/joshuadavidnelson/disable-blog
  * @since      0.4.0
- *
  * @package    Disable_Blog
- * @subpackage Disable_Blog/includes
+ * @subpackage Disable_Blog\Includes
+ * @author     Joshua Nelson <josh@joshuadnelson.com>
  */
 
 /**
@@ -16,9 +16,7 @@
  * the plugin, and register them with the WordPress API. Call the
  * run function to execute the list of actions and filters.
  *
- * @package    Disable_Blog
- * @subpackage Disable_Blog/includes
- * @author     Joshua Nelson <josh@joshuadnelson.com>
+ * @since 0.4.0
  */
 class Disable_Blog_Loader {
 
@@ -50,6 +48,25 @@ class Disable_Blog_Loader {
 		$this->actions = array();
 		$this->filters = array();
 
+	}
+
+	/**
+	 * Simple Autoloader.
+	 *
+	 * @since 0.5.1
+	 * @param string $class The Class to autoload.
+	 */
+	public function autoloader( $class ) {
+
+		$class   = 'class-' . str_replace( '_', '-', strtolower( $class ) ) . '.php';
+		$path    = plugin_dir_path( dirname( __FILE__ ) );
+		$sources = array( 'includes' );
+
+		foreach ( $sources as $source ) {
+			if ( file_exists( $path . $source . '/' . $class ) ) {
+				include $path . $source . '/' . $class;
+			}
+		}
 	}
 
 	/**
@@ -106,6 +123,56 @@ class Disable_Blog_Loader {
 
 		return $hooks;
 
+	}
+
+	/**
+	 * Remove a filter from the collection registered with WordPress.
+	 *
+	 * @since 0.5.1
+	 * @param string $tag              The filter hook to which the function to be removed is hooked.
+	 * @param string $class_name       Class name registering the filter callback.
+	 * @param string $method_to_remove Method name for the filter's callback.
+	 * @param int    $priority         The priority of the method (default 10).
+	 *
+	 * @return bool $removed Whether the function is removed.
+	 */
+	public function remove_filter( $tag, $class_name = '', $method_to_remove = '', $priority = 10 ) {
+
+		global $wp_filter;
+		$removed = false;
+
+		foreach ( $wp_filter[ $tag ]->callbacks as $filter_priority => $filters ) {
+
+			if ( $filter_priority == $priority ) {
+
+				foreach ( $filters as $filter ) {
+
+					if ( $filter['function'][1] == $method_to_remove
+						&& is_object( $filter['function'][0] ) // only WP 4.7 and above. This plugin is requiring at least WP 4.9.
+						&& $filter['function'][0] instanceof $class_name ) {
+						$removed = $wp_filter[ $tag ]->remove_filter( $tag, array( $filter['function'][0], $method_to_remove ), $priority );
+					}
+				}
+			}
+		}
+
+		return $removed;
+
+	}
+
+	/**
+	 * Remove an action from the collection registered with WordPress.
+	 *
+	 * @since 0.5.1
+	 * @param string $tag              The filter hook to which the function to be removed is hooked.
+	 * @param string $class_name       Class name registering the filter callback.
+	 * @param string $method_to_remove Method name for the filter's callback.
+	 * @param int    $priority         The priority of the method (default 10).
+	 *
+	 * @return bool $removed Whether the function is removed.
+	 */
+	public function remove_action( $tag, $class_name = '', $method_to_remove = '', $priority = 10 ) {
+		return $this->remove_filter( $tag, $class_name, $method_to_remove, $priority );
 	}
 
 	/**